+++ changelog.md
@@ -1,6 +1,7 @@
 ## Version 1.15.12+dev
  ### Add-ons client
    * The details panel now shows the list of tags in each add-on.
+   * Added a filter based on tags.
  ### Add-ons server
  ### Campaigns
  ### Editor